# [bounty] move "examples/hlb_cifar10.py" each epoch data processing from numpy into tinygrad (random_crop + cutmix)